refactor: 이벤트 핸들러위치 메인으로 이동
